Moved the date in education component into date.js; moved the isAfter function into date.js and also import it into generatecv
